+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release 1.14.0
 ## General
 - Added Global Mute toggle to Meadow Remix's "Gameplay" tab
+## Arena
+### Modders
+-  ⚠️ Updated `arena.ExternalGameMode.GetGamemodeId` to become a getter
+- Added `FinalResultBox_ctor` to `arena.ExternalGameMode`
+- Prevented Overseer from earning score
 
 # Release 1.13.2
 - Fixed Meadow crashing in single-player